fix(#3811): census finding C — nested UDT decoders enforce their own bounded-slice contract; 4 discarding callers of parse_udt_value now check
