+++ cqlite-core/src/storage/sstable/reader/parsing/row_decoder/cell_value_complex.rs
@@ -121,7 +121,11 @@ impl V5CompressedLegacyParser {
 
                 // Parse UDT value from the blob
                 let udt_data = &data[off..off + blob_len];
-                let (udt_value, _) = self.parse_udt_value(udt_data, 0, &udt_def, column)?;
+                let (udt_value, n) = self.parse_udt_value(udt_data, 0, &udt_def, column)?;
+                // #3811 (finding C): `parse_udt_value` REPORTS; this caller dropped it,
+                // so a frozen UDT blob with trailing bytes or a partial trailing field
+                // header was accepted where the collection and tuple paths refuse.
+                Self::require_fully_consumed_raw(n, udt_data.len(), &column.name, "frozen UDT")?;
                 off += blob_len;
 
                 (udt_value, off)
@@ -169,7 +173,11 @@ impl V5CompressedLegacyParser {
                 }
 
                 let udt_data = &data[off..off + blob_len];
-                let (udt_value, _) = self.parse_udt_value(udt_data, 0, &udt_def, column)?;
+                let (udt_value, n) = self.parse_udt_value(udt_data, 0, &udt_def, column)?;
+                // #3811 (finding C): `parse_udt_value` REPORTS; this caller dropped it,
+                // so a frozen UDT blob with trailing bytes or a partial trailing field
+                // header was accepted where the collection and tuple paths refuse.
+                Self::require_fully_consumed_raw(n, udt_data.len(), &column.name, "frozen UDT")?;
                 off += blob_len;
 
                 (udt_value, off)
